pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "captionalchemy"
version = "0.1.0"
description = "A Python package to create closed captions with face detection and recognition."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Ben Batman", email = "benbatman2@gmail.com" }]

dependencies = [
    "pyannote.audio==3.3.2",
    "pydub==0.25.1",
    "openai-whisper==20240930",
    "python-dotenv==1.1.0",
    "opencv-python==4.11.0.86",
    "numpy==2.1.3",
    "protobuf==5.29.4",
    "insightface==0.7.3",
    "onnxruntime==1.22.0",
    "panns-inference==0.1.1",
    "librosa==0.11.0",
    "tqdm==4.67.1",
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

requires-python = ">=3.8"

[project.scripts]
captionalchemy = "captionalchemy.caption:main"
